requirements.txt
langchain
langchain-community
langchain_openai
openai
python-dotenv
langchain-pinecone
langgraph
fitz
unstructured
pdfminer
pi_heif
unstructured_inference
pdf2image
google-cloud-vision
google-cloud-storage